+++ modules/anime/src/main/kotlin/com/thekeeperofpie/artistalleydatabase/anime/forum/thread/ForumThreadEntry.kt
@@ -6,6 +6,6 @@ import com.anilist.fragment.ForumThread
 data class ForumThreadEntry(
     val thread: ForumThread,
     val bodyMarkdown: Spanned?,
-    val liked: Boolean,
-    val subscribed: Boolean,
+    val liked: Boolean = thread.isLiked ?: false,
+    val subscribed: Boolean = thread.isSubscribed ?: false,
 )